added all files except bingo 7 final whose model is running
